fix(web): move Add Prospect out of the sidebar and onto the Queue
It was the second item in the sidebar — ahead of the Queue and the
Replies you read every day — for something you do occasionally. The ⌘K
palette had already reached the opposite conclusion: its "Go" group lists
Today, Queue, Cadences, Receipts, Measure, Plays and Setup, and omits Add
Prospect entirely. The sidebar was the only way in.

It becomes an action on the Queue instead, in the Target Queue caption
row, opening the same form in a modal. That is the page the result lands
on, and the empty state there already explains how candidates arrive — it
now names adding one by hand as the second path.

The form moves to components/queue/AddProspectForm.tsx so the page and
the modal cannot drift apart. Its one behavioural difference is driven by
`onQueued`: the modal closes on success, so the confirmation panel would
never be seen and is skipped — the toast and the new row say it landed.
The page has nowhere to go, so it keeps the panel and its link to the
Queue.

The route stays. Bookmarks resolve, the empty state can link to it, and
it is now in the palette where a keyboard user would look for it.
